+++ source/scone/ui/label.d
@@ -1,0 +1,11 @@
+module ui.label;
+
+import ui.element;
+
+class UILabel : UIElement
+{
+    this(string id, int x, int y, string text)
+    {
+        super(id, x, y, text);
+    }
+}