+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## v1.1.0
+_2 July 2016_
+
+ * Simplified initialization. renamed `sconeInit(SconeModule module)` to `sconeOpen()`.
+ * Added UI library, makes it easier for input boxes, such as name fields.
+ * Add localization support.
+
 ## v1.0.1
 _1 July 2016_